+++ tests/testthat/test-hop-index-vec.R
@@ -3,13 +3,24 @@
 
 test_that("size of each `.f` result must be 1", {
   expect_snapshot({
-    (expect_error(hop_index_vec(1, 1, 1, 1, ~c(.x, 1))))
+    (expect_error(hop_index_vec(1, 1, 1, 1, ~ c(.x, 1))))
   })
 })
 
 test_that("inner type is allowed to be different", {
   expect_equal(
-    hop_index_vec(1:2, 1:2, 1:2, 1:2, ~if (.x == 1L) {list(1)} else {list("hi")}, .ptype = list()),
+    hop_index_vec(
+      1:2,
+      1:2,
+      1:2,
+      1:2,
+      ~ if (.x == 1L) {
+        list(1)
+      } else {
+        list("hi")
+      },
+      .ptype = list()
+    ),
     list(1, "hi")
   )
 })
@@ -17,7 +28,18 @@ test_that("inner type is allowed to be different", {
 test_that("inner type can be restricted with list_of", {
   expect_snapshot({
     (expect_error(
-      hop_index_vec(1:2, 1:2, 1:2, 1:2, ~if (.x == 1L) {list_of(1)} else {list_of("hi")}, .ptype = list_of(.ptype = double())),
+      hop_index_vec(
+        1:2,
+        1:2,
+        1:2,
+        1:2,
+        ~ if (.x == 1L) {
+          list_of(1)
+        } else {
+          list_of("hi")
+        },
+        .ptype = list_of(.ptype = double())
+      ),
       class = "vctrs_error_incompatible_type"
     ))
   })
@@ -29,7 +51,10 @@ test_that("inner type can be restricted with list_of", {
 test_that(".ptype is respected", {
   expect_equal(hop_index_vec(1, 1, 1, 1, ~.x), 1)
   expect_equal(hop_index_vec(1, 1, 1, 1, ~.x, .ptype = int()), 1L)
-  expect_error(hop_index_vec(1, 1, 1, 1, ~.x + .5, .ptype = integer()), class = "vctrs_error_cast_lossy")
+  expect_error(
+    hop_index_vec(1, 1, 1, 1, ~ .x + .5, .ptype = integer()),
+    class = "vctrs_error_cast_lossy"
+  )
 })
 
 test_that("`.ptype = NULL` results in 'guessed' .ptype", {
@@ -42,7 +67,14 @@ test_that("`.ptype = NULL` results in 'guessed' .ptype", {
 test_that("`.ptype = NULL` fails if no common type is found", {
   expect_snapshot({
     (expect_error(
-      hop_index_vec(1:2, 1:2, 1:2, 1:2, ~ifelse(.x == 1L, "hello", 1), .ptype = NULL),
+      hop_index_vec(
+        1:2,
+        1:2,
+        1:2,
+        1:2,
+        ~ ifelse(.x == 1L, "hello", 1),
+        .ptype = NULL
+      ),
       class = "vctrs_error_incompatible_type"
     ))
   })
@@ -50,8 +82,34 @@ test_that("`.ptype = NULL` fails if no common type is found", {
 
 test_that("`.ptype = NULL` validates that element lengths are 1", {
   expect_snapshot({
-    (expect_error(hop_index_vec(1:2, 1:2, 1:2, 1:2, ~if(.x == 1L) {1:2} else {1}, .ptype = NULL)))
-    (expect_error(hop_index_vec(1:2, 1:2, 1:2, 1:2, ~if(.x == 1L) {NULL} else {2}, .ptype = NULL)))
+    (expect_error(
+      hop_index_vec(
+        1:2,
+        1:2,
+        1:2,
+        1:2,
+        ~ if (.x == 1L) {
+          1:2
+        } else {
+          1
+        },
+        .ptype = NULL
+      )
+    ))
+    (expect_error(
+      hop_index_vec(
+        1:2,
+        1:2,
+        1:2,
+        1:2,
+        ~ if (.x == 1L) {
+          NULL
+        } else {
+          2
+        },
+        .ptype = NULL
+      )
+    ))
   })
 })
 
@@ -89,7 +147,14 @@ test_that("`.ptype = NULL` errors with non recyclable starts/stops", {
 
 test_that(".ptypes with a vec_proxy() are restored to original type", {
   expect_s3_class(
-    hop_index_vec(Sys.Date() + 1:5, 1:5, 1:5, 1:5, ~.x, .ptype = as.POSIXlt(Sys.Date())),
+    hop_index_vec(
+      Sys.Date() + 1:5,
+      1:5,
+      1:5,
+      1:5,
+      ~.x,
+      .ptype = as.POSIXlt(Sys.Date())
+    ),
     "POSIXlt"
   )
 })
@@ -105,11 +170,20 @@ test_that("can return a matrix and rowwise bind the results together", {
 test_that("`hop_index_vec()` falls back to `c()` method as required", {
   local_c_foobar()
 
-  expect_identical(hop_index_vec(1:3, 1:3, 1:3, 1:3, ~foobar(.x), .ptype = foobar(integer())), foobar(1:3))
-  expect_condition(hop_index_vec(1:3, 1:3, 1:3, 1:3, ~foobar(.x), .ptype = foobar(integer())), class = "slider_c_foobar")
+  expect_identical(
+    hop_index_vec(1:3, 1:3, 1:3, 1:3, ~ foobar(.x), .ptype = foobar(integer())),
+    foobar(1:3)
+  )
+  expect_condition(
+    hop_index_vec(1:3, 1:3, 1:3, 1:3, ~ foobar(.x), .ptype = foobar(integer())),
+    class = "slider_c_foobar"
+  )
 
-  expect_identical(hop_index_vec(1:3, 1:3, 1:3, 1:3, ~foobar(.x)), foobar(1:3))
-  expect_condition(hop_index_vec(1:3, 1:3, 1:3, 1:3, ~foobar(.x)), class = "slider_c_foobar")
+  expect_identical(hop_index_vec(1:3, 1:3, 1:3, 1:3, ~ foobar(.x)), foobar(1:3))
+  expect_condition(
+    hop_index_vec(1:3, 1:3, 1:3, 1:3, ~ foobar(.x)),
+    class = "slider_c_foobar"
+  )
 })
 
 # ------------------------------------------------------------------------------
@@ -120,7 +194,7 @@ test_that("names exist on inner sliced elements", {
   x <- set_names(1:5, names)
   exp <- as.list(names)
   i <- vec_seq_along(x)
-  expect_equal(hop_index_vec(x, i, i, i, ~list(names(.x))), exp)
+  expect_equal(hop_index_vec(x, i, i, i, ~ list(names(.x))), exp)
 })
 
 test_that("names are never placed on the output", {