+++ tests/testthat/test-pslide.R
@@ -12,15 +12,25 @@ test_that("Recycling is carried out using tidyverse recycling rules", {
 
   expect_equal(pslide(list(x0, x0), ~.x), list())
   expect_equal(pslide(list(x0, x1), ~.x), list())
-  expect_snapshot((expect_error(pslide(list(x0, x2), ~.x), class = "vctrs_error_incompatible_size")))
+  expect_snapshot(
+    (expect_error(
+      pslide(list(x0, x2), ~.x),
+      class = "vctrs_error_incompatible_size"
+    ))
+  )
   expect_equal(pslide(list(x1, x1), ~.x), list(x1))
   expect_equal(pslide(list(x1, x2), ~.x), list(x1, x1))
-  expect_snapshot((expect_error(pslide(list(x2, x3), ~.x), class = "vctrs_error_incompatible_size")))
+  expect_snapshot(
+    (expect_error(
+      pslide(list(x2, x3), ~.x),
+      class = "vctrs_error_incompatible_size"
+    ))
+  )
 })
 
 test_that("pslide() can iterate over a data frame", {
   x <- data.frame(x = 1:5, y = 6:10)
-  expect_equal(pslide(x, ~.x + .y), as.list(x$x + x$y))
+  expect_equal(pslide(x, ~ .x + .y), as.list(x$x + x$y))
 })
 
 test_that("pslide() can iterate over a data frame with a data frame column", {
@@ -28,7 +38,7 @@ test_that("pslide() can iterate over a data frame with a data frame column", {
   x$x <- x
 
   expect_equal(
-    pslide(x, ~list(...)),
+    pslide(x, ~ list(...)),
     list(as.list(vec_slice(x, 1)), as.list(vec_slice(x, 2)))
   )
 })
@@ -38,7 +48,10 @@ test_that("pslide() requires a list-like input", {
 })
 
 test_that("pslide() forces arguments in the same way as base R / pmap()", {
-  f_slide <- pslide(list(1:2, 1:2, 1:2), function(i, j, k) function(x) x + i + j + k)
+  f_slide <- pslide(
+    list(1:2, 1:2, 1:2),
+    function(i, j, k) function(x) x + i + j + k
+  )
   f_base <- mapply(function(i, j, k) function(x) x + i + j + k, 1:2, 1:2, 1:2)
 
   expect_equal(f_slide[[1]](0), f_base[[1]](0))