+++ tests/testthat/test-slide-vec.R
@@ -2,16 +2,24 @@
 # type / size strict-ness
 
 test_that("size of each `.f` result must be 1", {
-  expect_snapshot(error = TRUE, slide_vec(1:2, ~c(.x, 1)))
+  expect_snapshot(error = TRUE, slide_vec(1:2, ~ c(.x, 1)))
 })
 
 test_that("size of each `.f` result must be 1", {
-  expect_snapshot(error = TRUE, slide_dbl(1:2, ~c(.x, 1)))
+  expect_snapshot(error = TRUE, slide_dbl(1:2, ~ c(.x, 1)))
 })
 
 test_that("inner type is allowed to be different", {
   expect_equal(
-    slide_vec(1:2, ~if (.x == 1L) {list(1)} else {list("hi")}, .ptype = list()),
+    slide_vec(
+      1:2,
+      ~ if (.x == 1L) {
+        list(1)
+      } else {
+        list("hi")
+      },
+      .ptype = list()
+    ),
     list(1, "hi")
   )
 })
@@ -19,7 +27,15 @@ test_that("inner type is allowed to be different", {
 test_that("inner type can be restricted with list_of", {
   expect_snapshot({
     (expect_error(
-      slide_vec(1:2, ~if (.x == 1L) {list_of(1)} else {list_of("hi")}, .ptype = list_of(.ptype = double())),
+      slide_vec(
+        1:2,
+        ~ if (.x == 1L) {
+          list_of(1)
+        } else {
+          list_of("hi")
+        },
+        .ptype = list_of(.ptype = double())
+      ),
       class = "vctrs_error_incompatible_type"
     ))
   })
@@ -28,7 +44,14 @@ test_that("inner type can be restricted with list_of", {
 test_that("inner type can be restricted", {
   expect_snapshot({
     (expect_error(
-      slide_dbl(1:2, ~if (.x == 1L) {1} else {"x"}),
+      slide_dbl(
+        1:2,
+        ~ if (.x == 1L) {
+          1
+        } else {
+          "x"
+        }
+      ),
       class = "vctrs_error_incompatible_type"
     ))
   })
@@ -41,7 +64,10 @@ test_that(".ptype is respected", {
   expect_equal(slide_vec(1, ~.x), 1)
   expect_equal(slide_vec(1, ~.x, .ptype = int()), 1L)
   expect_snapshot({
-    (expect_error(slide_vec(1, ~.x + .5, .ptype = integer()), class = "vctrs_error_cast_lossy"))
+    (expect_error(
+      slide_vec(1, ~ .x + .5, .ptype = integer()),
+      class = "vctrs_error_cast_lossy"
+    ))
   })
 })
 
@@ -55,7 +81,7 @@ test_that("`.ptype = NULL` results in 'guessed' .ptype", {
 test_that("`.ptype = NULL` fails if no common type is found", {
   expect_snapshot({
     (expect_error(
-      slide_vec(1:2, ~ifelse(.x == 1L, "hello", 1), .ptype = NULL),
+      slide_vec(1:2, ~ ifelse(.x == 1L, "hello", 1), .ptype = NULL),
       class = "vctrs_error_incompatible_type"
     ))
   })
@@ -63,7 +89,15 @@ test_that("`.ptype = NULL` fails if no common type is found", {
 
 test_that("`.ptype = NULL` validates that element lengths are 1", {
   expect_snapshot(error = TRUE, {
-    slide_vec(1:2, ~if(.x == 1L) {1:2} else {1}, .ptype = NULL)
+    slide_vec(
+      1:2,
+      ~ if (.x == 1L) {
+        1:2
+      } else {
+        1
+      },
+      .ptype = NULL
+    )
   })
 })
 
@@ -94,11 +128,17 @@ test_that("can return a matrix and rowwise bind the results together", {
 test_that("`slide_vec()` falls back to `c()` method as required", {
   local_c_foobar()
 
-  expect_identical(slide_vec(1:3, ~foobar(.x), .ptype = foobar(integer())), foobar(1:3))
-  expect_condition(slide_vec(1:3, ~foobar(.x), .ptype = foobar(integer())), class = "slider_c_foobar")
+  expect_identical(
+    slide_vec(1:3, ~ foobar(.x), .ptype = foobar(integer())),
+    foobar(1:3)
+  )
+  expect_condition(
+    slide_vec(1:3, ~ foobar(.x), .ptype = foobar(integer())),
+    class = "slider_c_foobar"
+  )
 
-  expect_identical(slide_vec(1:3, ~foobar(.x)), foobar(1:3))
-  expect_condition(slide_vec(1:3, ~foobar(.x)), class = "slider_c_foobar")
+  expect_identical(slide_vec(1:3, ~ foobar(.x)), foobar(1:3))
+  expect_condition(slide_vec(1:3, ~ foobar(.x)), class = "slider_c_foobar")
 })
 
 # ------------------------------------------------------------------------------
@@ -107,20 +147,41 @@ test_that("`slide_vec()` falls back to `c()` method as required", {
 test_that(".step produces typed `NA` values", {
   expect_identical(slide_int(1:3, identity, .step = 2), c(1L, NA, 3L))
   expect_identical(slide_dbl(1:3, identity, .step = 2), c(1, NA, 3))
-  expect_identical(slide_chr(c("a", "b", "c"), identity, .step = 2), c("a", NA, "c"))
+  expect_identical(
+    slide_chr(c("a", "b", "c"), identity, .step = 2),
+    c("a", NA, "c")
+  )
   expect_identical(slide_vec(1:3, identity, .step = 2), c(1L, NA, 3L))
-  expect_identical(slide_vec(1:3, identity, .step = 2, .ptype = integer()), c(1L, NA, 3L))
+  expect_identical(
+    slide_vec(1:3, identity, .step = 2, .ptype = integer()),
+    c(1L, NA, 3L)
+  )
 })
 
 # ------------------------------------------------------------------------------
 # .complete
 
 test_that(".complete produces typed `NA` values", {
-  expect_identical(slide_int(1:3, ~1L, .before = 1, .complete = TRUE), c(NA, 1L, 1L))
-  expect_identical(slide_dbl(1:3, ~1, .before = 1, .complete = TRUE), c(NA, 1, 1))
-  expect_identical(slide_chr(1:3, ~"1", .before = 1, .complete = TRUE), c(NA, "1", "1"))
-  expect_identical(slide_vec(1:3, ~1, .before = 1, .complete = TRUE), c(NA, 1, 1))
-  expect_identical(slide_vec(1:3, ~1, .before = 1, .complete = TRUE, .ptype = integer()), c(NA, 1L, 1L))
+  expect_identical(
+    slide_int(1:3, ~1L, .before = 1, .complete = TRUE),
+    c(NA, 1L, 1L)
+  )
+  expect_identical(
+    slide_dbl(1:3, ~1, .before = 1, .complete = TRUE),
+    c(NA, 1, 1)
+  )
+  expect_identical(
+    slide_chr(1:3, ~"1", .before = 1, .complete = TRUE),
+    c(NA, "1", "1")
+  )
+  expect_identical(
+    slide_vec(1:3, ~1, .before = 1, .complete = TRUE),
+    c(NA, 1, 1)
+  )
+  expect_identical(
+    slide_vec(1:3, ~1, .before = 1, .complete = TRUE, .ptype = integer()),
+    c(NA, 1L, 1L)
+  )
 })
 
 # ------------------------------------------------------------------------------
@@ -130,7 +191,7 @@ test_that("names exist on inner sliced elements", {
   names <- letters[1:5]
   x <- set_names(1:5, names)
   exp <- set_names(as.list(names), names)
-  expect_equal(slide_vec(x, ~list(names(.x))), exp)
+  expect_equal(slide_vec(x, ~ list(names(.x))), exp)
 })
 
 test_that("names can be placed on atomics", {
@@ -146,28 +207,32 @@ test_that("names can be placed on atomics", {
 test_that("names from `.x` are kept, and new names from `.f` results are dropped", {
   x <- set_names(1, "x")
 
-  expect_identical(slide_vec(x, ~c(y = 2), .ptype = NULL), c(x = 2))
-  expect_identical(slide_vec(1, ~c(y = 2), .ptype = NULL), 2)
+  expect_identical(slide_vec(x, ~ c(y = 2), .ptype = NULL), c(x = 2))
+  expect_identical(slide_vec(1, ~ c(y = 2), .ptype = NULL), 2)
 
-  expect_identical(slide_dbl(x, ~c(y = 2)), c(x = 2))
-  expect_identical(slide_dbl(1, ~c(y = 2)), 2)
+  expect_identical(slide_dbl(x, ~ c(y = 2)), c(x = 2))
+  expect_identical(slide_dbl(1, ~ c(y = 2)), 2)
 })
 
 test_that("names can be placed on data frames", {
   names <- letters[1:2]
   x <- set_names(1:2, names)
 
-  out <- slide_vec(x, ~data.frame(x = .x))
+  out <- slide_vec(x, ~ data.frame(x = .x))
   expect_equal(rownames(out), names)
 
-  out <- slide_vec(x, ~data.frame(x = .x), .ptype = data.frame(x = int()))
+  out <- slide_vec(x, ~ data.frame(x = .x), .ptype = data.frame(x = int()))
   expect_equal(rownames(out), names)
 })
 
 test_that("names can be placed on arrays", {
   names <- letters[1:2]
   x <- set_names(1:2, names)
-  out <- slide_vec(x, ~array(.x, c(1, 1)), .ptype = array(int(), dim = c(0, 1)))
+  out <- slide_vec(
+    x,
+    ~ array(.x, c(1, 1)),
+    .ptype = array(int(), dim = c(0, 1))
+  )
   expect_equal(rownames(out), names)
 })
 
@@ -223,7 +288,7 @@ test_that("slide_dfr() works", {
   expect_identical(
     slide_dfr(
       1:2,
-      ~new_data_frame(list(x = list(.x))),
+      ~ new_data_frame(list(x = list(.x))),
       .before = 1
     ),
     data_frame(