+++ NEWS.md
@@ -1,3 +1,9 @@
+# slider 0.3.3
+
+* Fixed the C level function signature for a vctrs callable (#224).
+
+* Removed usage of non-API `OBJECT()`.
+
 # slider 0.3.2
 
 * Removed usage of non-API `STRING_PTR()` (#209).